Add more samplers.
For completeness' sake, I've added the rest of the samplers described in
the book. None of them are really tested at all, and there are
definitely improvements to be made still, but now I can move on with
confidence!

Incidentally, I may add to this list of samplers over time, with other
well-known 2D distributions with good behaviour. This book is rather old
after all.