Split up into separate binary and library crates.
This allows me to remove the large numbers of `#![allow(dead_code)]`
that were springing up for things I didn't happen to use in my test
scenes.

It's also a nicer separation of concerns, and will make it easier for me
to maintain multiple executables for e.g. different scenes that I want
to experiment with and/or demonstrate simultaenously, without having to
have ugly commented-out code or something.